+++ plexsvc-framework/src/main/java/com/plexobject/handler/Response.java
@@ -17,7 +17,7 @@ public class Response extends BasePayload<Object> {
     public static final String[] HEADER_PROPERTIES = new String[] {
             HttpResponse.STATUS_CODE, HttpResponse.STATUS_MESSAGE,
             HttpResponse.LOCATION, Constants.SESSION_ID,
-            HttpResponse.VERSION_HEADER };
+            HttpResponse.VERSION_HEADER, HttpResponse.RESPONSE_MILLITIME_HEADER };
     private final Request request;
 
     public Response(final Request request,
@@ -51,6 +51,12 @@ public class Response extends BasePayload<Object> {
 
     public Response setVersionHeader(String version) {
         properties.put(HttpResponse.VERSION_HEADER, version);
+        return this;
+    }
+
+    public Response setResponseMilliTime(long millisTime) {
+        properties.put(HttpResponse.RESPONSE_MILLITIME_HEADER,
+                String.valueOf(millisTime));
         return this;
     }