+++ plexsvc-framework/src/main/java/com/plexobject/http/HttpResponse.java
@@ -11,6 +11,7 @@ public interface HttpResponse {
     public static final String LOCATION = "location";
     public static final String ERROR_MESSAGE = "errorMessage";
     public static final String VERSION_HEADER = "X-API-Version";
+    public static final String RESPONSE_MILLITIME_HEADER = "X-Response-MilliTime";
 
     /**
      * Status code (100) indicating the client can continue.